app/requirements.txt
flask
sqlalchemy
requests
numpy
pandas
PyJWT==1.7.1